requirements.txt
Flask
gunicorn
flask-sqlalchemy
pymysql
flask-admin
flask-login
Pillow